EE-993: instrument Wasm instructions
This commit adds instrumentation for individual Wasm instructions.  It should
give a notion of relative execution times for each instruction, allowing for
more accurate allocation of gas costs in
https://github.com/paritytech/wasm-utils/blob/master/src/rules.rs

The code should not be merged to the production codebase as is.  It would
require the instrumentation logic to be e.g. feature-gated so that it doesn't
affect the normal running of the crate.

Only non-floating-point instructions generate instrumentation since our current
usage forbids these.  It is trivial to change to include instrumentation of
floating point instructions though.

The results are produced as a file per instruction in 'target/metrics' and are
created by running:
cargo test --release spec -- --test-threads=1

The results files are appended to on each run, and are set to cap their contents
at 10,000 entries.  In this way, the test suite can be run in a loop in order to
generate sufficient entries in each file.